refactor(plugins/k8saudit): create package directory
