+++ plugins/k8saudit/.gitignore
@@ -1,1 +1,1 @@
-test_files/*
+test_files
