+++ plugins/k8saudit/pkg/k8saudit/extract_test.go
@@ -1,3 +1,19 @@
+/*
+Copyright (C) 2022 The Falco Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
 package k8saudit
 
 import (
@@ -107,8 +123,8 @@ func readTestFiles(b testing.TB) []string {
 
 func BenchmarkExtractFromJSON(b *testing.B) {
 	req := &testExtractRequest{}
-	e := &AuditEventExtractor{}
-	fields := Fields()
+	e := &Plugin{}
+	fields := e.Fields()
 	jsons := readTestFiles(b)
 
 	b.ResetTimer()
@@ -118,7 +134,7 @@ func BenchmarkExtractFromJSON(b *testing.B) {
 		for ev, data := range jsons {
 			for f, field := range fields {
 				fieldEntryToRequest(uint64(f), &field, req)
-				json, err := e.Decode(uint64(ev), strings.NewReader(data))
+				json, err := e.DecodeReader(uint64(ev), strings.NewReader(data))
 				if err != nil && err != ErrExtractNotAvailable {
 					b.Errorf("decoding field %s: %s", field.Name, err.Error())
 				}
